# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Реактив CRM",
    "description": "CRM продаж и закупок химической продукции. Настраиваемые финансовые профили, права и шаблоны. Количества и деньги — десятичные строки.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Login Api V1 Auth Login Post"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Me Api V1 Auth Me Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Logout",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object",
                  "title": "Response Logout Api V1 Auth Logout Post"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/auth/mfa/setup": {
      "post": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Mfa Setup",
        "operationId": "mfa_setup_api_v1_auth_mfa_setup_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Mfa Setup Api V1 Auth Mfa Setup Post"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/auth/mfa/enable": {
      "post": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Mfa Enable",
        "operationId": "mfa_enable_api_v1_auth_mfa_enable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object",
                  "title": "Response Mfa Enable Api V1 Auth Mfa Enable Post"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/auth/password": {
      "post": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Change Password",
        "operationId": "change_password_api_v1_auth_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object",
                  "title": "Response Change Password Api V1 Auth Password Post"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Directory",
        "operationId": "directory_api_v1_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Directory Api V1 Users Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/admin/permissions": {
      "get": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Permissions",
        "operationId": "permissions_api_v1_admin_permissions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Permissions Api V1 Admin Permissions Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Users",
        "operationId": "users_api_v1_admin_users_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Users Api V1 Admin Users Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      },
      "post": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Create User",
        "operationId": "create_user_api_v1_admin_users_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create User Api V1 Admin Users Post"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/admin/users/{entity_id}": {
      "patch": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Edit User",
        "operationId": "edit_user_api_v1_admin_users__entity_id__patch",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Edit User Api V1 Admin Users  Entity Id  Patch"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/admin/roles": {
      "get": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Roles",
        "operationId": "roles_api_v1_admin_roles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Roles Api V1 Admin Roles Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      },
      "post": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Create Role",
        "operationId": "create_role_api_v1_admin_roles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Role Api V1 Admin Roles Post"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/admin/roles/{entity_id}": {
      "put": {
        "tags": [
          "Доступ и настройки"
        ],
        "summary": "Edit Role",
        "operationId": "edit_role_api_v1_admin_roles__entity_id__put",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Edit Role Api V1 Admin Roles  Entity Id  Put"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginInput": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Password"
          },
          "otp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Otp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginInput"
      },
...
      "OTPInput": {
        "properties": {
          "otp": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "Otp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "otp"
        ],
        "title": "OTPInput"
      },
      "PasswordInput": {
        "properties": {
          "current_password": {
            "type": "string",
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 12,
            "title": "New Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordInput"
      },
...
      "UserInput": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 12,
            "title": "Password"
          },
          "role_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Role Ids",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "name",
          "password"
        ],
        "title": "UserInput"
      },
...
    }
  }
}